refactor: clean up details screens (#77)
